tsc: a 3.9-billion count does not fit in an int
tscGetCyclesPerSecond() measured a uint64_t cycle count and returned it
as int: every CPU faster than ~2.1GHz overflowed int32, wrapped
negative, and sign-extended into kCPUCyclesPerSecond as garbage. On the
3900X that's ~3.9e9 cycles masquerading as a number near 2^64. Every
consumer of that global has been doing arithmetic with nonsense on real
hardware since the day it was written — found when CPU-time accounting
became the first consumer that checked its math.

+++ kernel/include/driver/system/x86_64.h
@@ -8,5 +8,5 @@
 	uint32_t read_apic_id();
 	uint64_t rdtsc();
 	uint64_t getCR3();
-	int tscGetCyclesPerSecond();
+	uint64_t tscGetCyclesPerSecond();
 #endif
